requirements.txt
streamlit==1.38.0
onnxruntime==1.22.0
pillow==10.4.0
numpy==2.1.1
